Исправления после ревью

+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -6,9 +6,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class Review {
-    private Long reviewId;
-    private Long userId;
-    private Long filmId;
+    private Integer reviewId;
+    private Integer userId;
+    private Integer filmId;
     private String content;
     private Boolean isPositive;
     private Integer useful; // рейтинг полезности (лайк или дизлайк)
